Issue #13 - Organization of "Launch Task" drop-down menu (!)
The menu has been replaced by a tiered jQuery UI one featuring the name of
the tasks' plugins, which are extracted out of the source file paths
given by the source location of plugin class methods. The fileless launch
mechanism (_launch_task.js.erb) has been updated accordingly and slightly
improved (it now shows the task's name in the buttons).

It has been necessary to update jQuery UI to 1.11.4 in order to
implement the menu. While is it backwards compatible with 1.8.3, further UI
testing is required to ensure everything still works properly with the
new version.