Issue #143 - Help user with file management
This changeset implements the solution suggested in the issue
description; project, data provider and bourreau comboboxes now have
item descriptions in tooltips (using the title HTML attribute to pass
the description to the client).

A refactor of the current client-side tooltip handler was required,
as the old version relied on relative positioning and having a valid
parent element for tooltip targets (the new version uses absolute
positioning and works no matter where the target or tooltip is located
in the DOM).